+++ gamevault/UserControls/InstallUserControl.xaml.cs
@@ -300,8 +300,11 @@ namespace gamevault.UserControls
 
             string path = "";
             KeyValuePair<Game, string> result = InstallViewModel.Instance.InstalledGames.Where(g => g.Key.ID == gameId).FirstOrDefault();
-            MainWindowViewModel.Instance.AppBarText = $"Syncing cloud save...";
-            await SaveGameHelper.Instance.RestoreBackup(gameId, result.Value);
+            if (SettingsViewModel.Instance.CloudSaves)
+            {
+                MainWindowViewModel.Instance.AppBarText = $"Syncing cloud save...";
+                await SaveGameHelper.Instance.RestoreBackup(gameId, result.Value);
+            }
 
             if (!result.Equals(default(KeyValuePair<Game, string>)))
             {