+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # GameVault App Changelog
 
+## 1.16.1
+Recommended Gamevault Server Version: `v14.1.0`
+### Changes
+
+- Added a setting to mount ISO instead of extracting
+- Bug fix: Custom Cloud Save config is not applied to 'find' action
+- Bug fix: Cloud Save Backup/Restore not working on windows username with space in it
+
 ## 1.16.0
 Recommended Gamevault Server Version: `v14.1.0`
 ### Changes